Fix wallet transaction creation for pre-CLSAG hardforks
The wallet was incorrectly using CLSAG signature type (RCTTypeCLSAG)
for all transactions regardless of the current hardfork version. This
caused transactions to be rejected by the daemon on networks running
hardfork versions before HF13 where CLSAG is not yet enabled.

The fix properly checks whether CLSAG is enabled before selecting the
RingCT signature type:
- HF15+: Use RCTTypeBulletproofPlus (bp_version 4)
- HF13-14: Use RCTTypeCLSAG (bp_version 3)
- HF10-12: Use RCTTypeBulletproof2 (bp_version 2)

This allows transactions to be created and accepted on all hardfork
versions.